refactor: 클린 아키텍처 구조에 따른 구조 개편
