+++ lib/data/repository/recipe_repository.dart
@@ -1,4 +1,4 @@
-import '../model/recipe.dart';
+import 'package:recipe_app/domain/model/recipe.dart';
 
 abstract interface class RecipeRepository {
   Future<List<Recipe>> getRecipes();
