+++ lib/domain/model/ingredient_with_amount.dart
