+++ lib/domain/model/user.dart
