+++ lib/presentation/component/ingredient_item.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:recipe_app/core/ui/color_styles.dart';
 import 'package:recipe_app/core/ui/text_styles.dart';
-import 'package:recipe_app/data/model/ingredient.dart';
+import 'package:recipe_app/domain/model/ingredient.dart';
 
 class IngredientItem extends StatelessWidget {
   final Ingredient ingredient;
