+++ lib/presentation/ingredient/search_recipes/search_recipes_view_model.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:recipe_app/core/enums/time_filter.dart';
-import 'package:recipe_app/data/model/recipe.dart';
 import 'package:recipe_app/data/repository/recipe_repository.dart';
 import 'package:recipe_app/core/enums/category_filter.dart';
+import 'package:recipe_app/domain/model/recipe.dart';
 import 'package:recipe_app/presentation/component/filter_search_state.dart';
 import 'package:recipe_app/presentation/ingredient/search_recipes/search_recipes_state.dart';
